More refactoring, more tests
